organization: move FEState into own header, move variant coef into top-level serac namespace
