+++ src/common/common.hpp
@@ -14,6 +14,7 @@
 #define COMMON
 
 #include "common/boundary_condition.hpp"
+#include "common/finite_element_state.hpp"
 #include "common/logger.hpp"
 #include "common/serac_types.hpp"
 #include "common/terminator.hpp"
