+++ src/solvers/elasticity_solver.cpp
@@ -134,7 +134,7 @@ void ElasticitySolver::QuasiStaticSolve()
   for (auto& bc : ess_bdr_) {
     bool should_be_scalar = false;
     bc.projectBdr(displacement_->gridFunc(), time_, should_be_scalar);
-    displacement_->gridFunc().GetTrueDofs(displacement_->trueVec());
+    displacement_->initializeTrueVec();
     mfem::EliminateBC(*K_mat_, *K_e_mat_, bc.getTrueDofs(), displacement_->trueVec(), *bc_rhs_);
   }
